new-sdk-solution/backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5
django-environ>=0.10
django-cors-headers>=3.14
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.2
whitenoise>=6.5
dj-rest-auth>=4.0.0
django-allauth>=0.60 